Preserve dist/ directory between builds
Previously this was removed when calling 'python build.py', which
would cause any existing server to crash (the inode it referred to
having been destroyed). It also conflicts with an upcomign commit
where we need to read existing files in dist/. This preserves the
directory between builds, and introduces a Makefile with a "clean"
rule to compensate for the original functionality, which is only
necessary for local (stateful) machine builds.