Introduce fingerprint on stylesheet asset
This mirrors the approach we take in the admin repo [1], but with
less code as we don't need to override part of Flask ("get_url").
When we upload the assets to S3, we'll set them as immutable, with
a very long cache time [2]; a change in the fingerprint will force
a cache miss and fetch the new version of the file.

[1]: https://github.com/alphagov/notifications-admin/blob/df7a7dba79bec3b00baf761ef072b37d473dc34c/app/asset_fingerprinter.py#L39
[2]: https://github.com/alphagov/notifications-admin/blob/df7a7dba79bec3b00baf761ef072b37d473dc34c/Makefile#L179