Add basic HTTP2 server using NodeJS
NodeJS has a `http2` module as part of its core
library. This uses it to serve the static pages
over HTTP locally.

It's mostly copied from: https://nodejs.org/api/http2.html

Note: HTTP2 needs to be served over HTTPS so this
also includes a certificate and private key.
Viewed in a browser, this will warn you that the
page is insecure so you'll need to confirm you
still want to proceed before the browser will show
the page.