Add static image for map
We have a few options for how to show the areas an
alert should reach, long-term:
1. use leafletJS, as we do in the admin interface,
2. use a static image, like this one but generated
   from alert data

Number 1. depends on our ability to find a way of
providing the tiles that complies with all related
licencing and can take the load we expect. Current
ideas are to use the Ordinance Survey data tile
service: https://osdatahub.os.uk/

Best option for number 2. seems to be:

https://github.com/flopp/py-staticmaps

In the meantime, we can at least add an example of
what number 2. will look like, which is the most
MVP of the options.