Make sharing button use Web Share API by default
The Web Share API is used to make sharing links
and other content easy between apps on devices:

https://w3c.github.io/web-share/

This is implemented in some browsers, on some
devices so we should use it where we can:

https://caniuse.com/web-share

This code is based on feature detection so will
fall back to letting you copy the URL to the
clipboard if Web Share is not supported.

I also wrapped the copy-to-clipboard code in
old-style conditional comments to stop them being
served to IE8 and below. The DOM APIs it uses were
added in IE9.