Make more information link work out of context
Makes it read like:
"More information about this alert to England,
Scotland and Wales"

Once people are able to use the admin interface to
create alerts this won't suffice as we could
easily have multiple alerts to the same areas.
This is a known problem though and the team are
looking into how each alert can usefully be
identified. We will look at this again when we get
to that stage.