Compress JS using Terser
Terser is a replacement for `uglify-es`, a version
of UglifyJS that can work with ES Modules.

Reasons for choosing Terser over alternatives

There don't seem to be many code minifiers for
rollup:

https://github.com/rollup/awesome#output

Of those, I looked at:
- UglifyJS
- Terser
- Google closure compiler

Terser seems to be a more future proofed version
of UglifyJS (a direct fork actually). It also
seems to have consistent contributions and decent
backing:
- https://github.com/terser/terser/graphs/contributors
- https://opencollective.com/terser

Google closure compiler has, apparently, the most
potential for optimisation but without any
configuration (which is said to be harder than
other minifiers), its output was larger than
Terser. Its backing and contributions are
obviously good as Google has a vested interest.

In all, I chose Terser because it shaved more off
file sizes with no tweaking, is a fork of
UglifyJS, which is used elsewhere in GDS and seems
to have good support.