Remake example image with HTML/CSS
Remakes the example image in HTML and CSS to
save bytes. Also puts the CSS inline as part of
that. Credit to @quis for coming up with the idea.

Adds 2.4KB to the HTML when delivered
unoptimised, about 0.68KB compressed.

It removes the need for an image that, when
compressed is 29KB.

Also delivers a fallback image for IE8 (which
doesn't support enough CSS to make the image in
CSS).

I checked on all the major screen readers and it
gets recognised correctly and the label read out
as image alt text.