Fix SVG fallback images
The GOV.UK one:

The path for the fallback image (the logo is SVG
so this is for browsers that don't support it) was
defaulting to `/assets/images/` which doesn't
match our paths.

The alerts SVG fallback image:

The SVG is positioned absolutely so the image it
contains needed dimensions in its CSS.

I also made a few tweaks to our SVG code, based on:

https://github.com/alphagov/govuk-frontend/blob/master/src/govuk/components/header/template.njk#L32

Both were missed due to lack of testing in IE8.