Add script to make dates relative and up-to-date
Uses a version of the library we use on the admin
app to make sure dates are displayed relative to
the current time.

For example if it is 15:00 on 1 March 2021, the
date/time of:

12:00 on 1 March 2021

...would display as:

'3 hours ago'

The script is marked as 'async' to prevent it
effecting the critical rendering path:

See
https://developers.google.com/web/fundamentals/performance/critical-rendering-path/analyzing-crp
for details.