Remake example image in SVG
Follows the approach to remake the image currently
made out of HTML and CSS in SVG while keeping the
semantics of a native <img> element:

https://www.scottohara.me/blog/2019/05/22/contextual-images-svgs-and-a11y.html#testing-the-svg-element

The changes to our code include:
1. removing the inline styles for the HTML version
2. moving the macro.html into example.html
3. moving the alt text into the component
4. adding some CSS for it into the main stylesheet

Removing the inline styles for the HTML version

These aren't needed as we control the SVG visuals
directly on their tags.

Moving the macro.html into example.html

The example component doesn't need to be split
into two any more now styles.html is removed.

Moving the alt text into the component

The component is only used with one variant of alt
text and moving it inside lets us reference the
popup text directly.

Adding some CSS for it into the main stylesheet

This CSS only positions it horizontally so putting
it inline seems unneeded when that's the only need
for it and users won't notice