Change popup font to Arial and adjust sizes
Apple's system fonts are the only ones that look
the right size for the body text of the popup and
fit on the lines horizontally.

Both the Android and Windows fonts display larger
and overrun their lines.

This sets the text to Arial instead and lowers the
size so it is as close to the design (which I
think uses the Apple font) as possible while
looking OK across OS's.