Use same font family on all devices
We tried to use a nicer font on Apple devices, but because the font had
different metrics we couldn’t get the spacing to match when reflowing
the text by hand.

This commit also reduces the font size by 1px, to account for Arial
being a bit wider than San Francisco, and then adjusts the spacing and
size of the icon accordingly.