Start JS tests with test for relative dates
Sets up the JS tests by:
- creating ./tests/javascripts
- installing the jest test framework to match the
  admin app
- adding an npm script for `npm test` that runs
  jest
- adds the rollup-jest plugin to transform our JS
  files, which are written as ECMAScript Modules
  (ESM), into JS that can be used by jest*
- adds tests for the relative-dates.js code

*jest runs in NodeJS and uses its `vm` API.
Support for ESM in that API is still flagged as
experimental.

Changes to JS files to allow testing

All of our source JS files are modules but some
don't export anything, they're just run for their
side effects (meaning the code is evaluated and
run when they are imported but they export
nothing).

This is fine for the design of modules as source
code for browser-bundles but causes problems when
trying to test this code. This is because modules
are evaluated once by default, at the point
they're imported, so don't really fit into test
runs (which involve things being set up and torn
down multiple times).

The most sensible solution seems to make these
modules export a function containing the code to
be run and have a module that runs them which is
only used for bundling. This is what these changes
do.

This commit also changes all JS coded as an
ECMAScript module to the `.mjs` extension, to make
them easier to identify and consistent with the
criteria NodeJS uses:

https://nodejs.org/dist/latest-v14.x/docs/api/packages.html#packages_determining_module_system