Specify npm version to be lower than 7.0.0
This is because fron npm version 7.0.0 the way that
package.lock.json is built changes. We do not want to have
discrepancies, so we are sticking with older versions
for now.

This should be reviewed at some point, when the latest
stable version of node comes with npm 7.0.0 or newer.

More details:
https://docs.npmjs.com/cli/v7/configuring-npm/package-lock-json#lockfileversion