Add test for inline script
We use the GOVUK Frontend library template.html
file as the base layout of our pages. It includes
an inline script which our Content Security Policy
allows by referencing a sha256 of its contents.

This adds a test to check the script's contents
match the sha256. If this test fails, the sha256
has changed and so the sha256 in our Content
Security Policy will need to be updated to the new
value.

The Content Security Policy is added here:

https://github.com/alphagov/notifications-broadcasts-infra/blob/main/terraform/modules/govuk-alerts-website/files/main.py

It is tested here:

https://github.com/alphagov/notifications-broadcasts-infra/blob/main/terraform/modules/govuk-alerts-website/files/test_inject_headers.py

Both contain references to the sha256.