Make 'bootstrap' task us `npm ci` & explain
In discussions on the associated pull request it
was mentioned that:
- 'bootstrap' should be something you do
  initially, not really for any subsequent changes
  to dependencies
- `npm ci` should be run initially in CI and
  locally

Because of this, it makes sense to just change the
'bootstrap' task to use `npm ci`.

It was also mentioned that in NodeJS its common to
update dependencies using `npm install` so this
also adds something to the README to explain this.