Add free stock images
Replacing the prototype images with ones from free stock sites so we
have the rights to use them.

We’ve been advised not to use imagery of emergency services personnel.

This shouldn’t be merged until these images are signed off.

Cutting down to three images because I couldn’t think of a good fourth
example.

Using a JPEG image for the non-WebP fallback because JPEF compresses
photos better.