Add survey component to pages
Based on the GOVUK survey component but stripped
down as the content for ours is just a paragraph.

The other difference is that the GOVUK version is
located after the `<main>` element whereas ours is
inside it at the end. The GOVUK Frontend template
we use for our base layout doesn't have a block
after `<main>`:

https://github.com/LandRegistry/govuk-frontend-jinja/blob/main/govuk_frontend_jinja/templates/template.html#L45

Because of this, we would have to copy the `main`
block from that template and rewrite it in our
layout to have a block after `main`. This would
mean we would then have to keep that up to date
with the version we copied.

I don't think that extra maintenance is worth the
semantic value we get from it sitting outside
`main`.