Fix converted dates working with is_current_alert
Previously we assumed is_current_alert would be called before the
timestamps were converted in place. This adapts the function to
work on the converted dates instead, which means we can call it
multiple times, to support previous alerts.

Note that the previous "as_datetime" property was never used, so
we can freely repurpose it for actual use.