Add details of planned test in Reading
So that this page doesn’t become outdated this commit replaces details
of the East Suffolk test (which will be on the previous alerts page)
with details of the upcoming test in Reading.

This test has been publicised in the media[1], so it’s OK for us to give
the date publicly. We can add more detail closer to the day.

---

1. https://www.bbc.co.uk/news/uk-politics-57145675