Add expired date to past alerts
The content on past alert pages needs to make clear that the alert is
not about a current emergency.

All we can say for certain is that the alert has stopped sending. The
emergency itself may be ongoing.

We agreed to prioritise the date and time over the body of the alert,
reversing the order of priority of information on current alerts.

---

Discussion: https://docs.google.com/document/d/1regPI_sAYQsHxUxgjQ3_nLiQ_E_qQkKiG4AIKATD0p8/edit#