Remove JS for GOVUK buttom component
We only have one button on our pages, the one
added to share/copy the URL for an alert.

The GOVUK button JS does two things:
- it polyfills the component variant rendered as a
  link so it can respond to a spacebar press
- it stops subsequent presses of the button
  working to stop forms submitting multiple times

We don't use the link variant of the button
component and pressing either the share or copy
versions of our button more than once causes no
issues* so we don't need this code.

*Copy-to-clipboard commands are idempotent and
sharing brings up a dialog which overlays the
button, preventing further clicks.