Refactor how GOVUK Frontend JS is included
The way we currently include the GOVUK Frontend JS
is to put all the component JS in a single file,
along with a function that matches DOM nodes to
component code. The structure of this file, and
its comments, are based on the one GOVUK Frontend
uses to include all the component code.

Our needs are slightly different. We need to:
- include the JS for the GOVUK details component
  in its own file
- the sharing button JS to be able to include the
  JS for the GOVUK button component in its file

This therefore proposes to change how we include
the GOVUK Frontend JS to:
- keep the JS for the GOVUK details component in
  its own file and make its format match the other
  modules we have (splitting its core code out
  from the code that runs it)
- import the GOVUK button component JS into the
  sharing button JS

I think that this new structure will result in us
having the files we need while providing examples
of the different ways you can include GOVUK
Frontend component JS.