Add JS for details component to /opt-out page
This was missed when we added the govukDetails
component to this page.

This is very easy to do because the HTML
govukDetails produces seems like it works in most
modern browsers. I've included a task to the story
adding tests to make sure we have a test that
checks for this in future:

https://www.pivotaltracker.com/story/show/177597156