Change hashing method and output non-hashed copy
Changes the plugin used to add a SHA to the
filenames of all file types except JavaScript.

For JavaScript files, use Rollup's own hashing
method. Note: this method means the output file
matches that input so references to it in
templates need to change.

This commit also creates a copy of each file it
adds a SHA to, so the python filter code can find
it without needing to know the SHA at any point.