Rewrite main.py to add Flask
This will remove the previous functionality that we could display the
alerts locally. Instead it adds a single endpoint `refresh-alerts` that
will build, upload to s3 and purge the fastly cache in a single call.

For now, for the purposes of this spike, the content of the alerts comes
from the `data.yml` file, but we should implement a similar endpoint in
notify-api to provide such data.