Add explainer film to the homepage
We’ve agreed with the comms folk that this video should go in the empty
space on the homepage.

I’ve borrowed the responsive embed CSS from the admin app to make sure
the video displays properly on different device sizes.

Hosted on Youtube, but served via their cookie-free domain so we don’t
need to add a cookie policy.