Return names of JS files to old format
It turns out that rollup's 'input' option allows
you to define the 'name' property (used in the
format string that defines the final name of
output files). This can be done by passing in an
object containing name:path pairs.

For example:

'input': {
  'foo': '/path/to/file.js'
},
'output': {
  'entryFileNames': '[name].js'
}

...will produce 'foo.js'

These changes use this feature to return the
JavaScript files to the names they had before and
updates the HTML to reflect this.

See the docs for the input option for more detail:

https://www.rollupjs.org/guide/en/#input