Test details component is used with its JS
We use the GOVUK Frontend details component when
we want to use the <details> tag in our HTML. To
do this, you need to include its Jinja macro and
its JavaScript in your Jinja template and its Sass
in main.scss. The Jinja provides the HTML and the
Sass its visuals. The JavaScript:
1. provides a polyfill to make <details> work in
   browsers that don't support it
2. adds ARIA atrributes and event handlers that
   ensure it communicates the right information to
   the OS's accessibility API (and so to assistive
   technologies) and works correctly for keyboard
   users

If you're just testing the visual interace works,
in a modern browser, you can use the Jinja macro
and Sass and it seems to all be working. It's only
if you use assistive technology or the keyboard
that you notice the difference made by not
including the JavaScript.

Because of this, we've accidently forgotten to
include the JavaScript when using the component in
pages, in this and other apps.

This adds a test to make sure you can't use the
component without including its JavaScript, to
prevent that scenario happening.

Note: full details of the functionality the
JavaScript adds can be read in its comments:

https://github.com/alphagov/govuk-frontend/blob/main/src/govuk/components/details/details.js