Add function to restore local development
The previous approach had to hack around the filesystem limitations,
that would not allow having both a file and a directory sharing the same
name (i.e. "alerts")

This takes a more dynamic approach where it reuses the existing dict
which contains the rendered content and serves it to the user.

This is only available in the development environment and it relies on
path manipulation to serve static content from flask.