Update all NPM modules to latest versions
Brings in major version bumps to:

- @rollup/plugin-commonjs
- @rollup/plugin-node-resolve
- govuk-frontend
- gulp-sass
- jest

Of those, the following require changes to our
code.

Gulp-sass

The gulp-sass API has changed, requiring you to
now set the compiler used in a different way.

The compiler we use, dart-sass, used the
node-fibers module to speed up rendering. Node
itself contains changes from version 16 and up
that mean it can no longer be used.

See this for more detail on node fibers'
deprecation: https://sass-lang.com/blog/node-fibers-discontinued

See this for more detail on the changes needed for
gulp-sass version 5: https://github.com/dlmanning/gulp-sass/tree/master#migrating-to-version-5

Sass

This pins the sass module to v1.32.13 because
versions after that produce warnings from the
GOVUK Frontend Sass code that it looks like won't
be changed for a while (see
alphagov/govuk-frontend#2238).

Jest

Version 27.0.0 of Jest changed the default test
environment from JSDOM (a JS module that simulates
the Document Object Model - DOM) to NodeJS.

Our tests rely on the environment being JSDOM so
we now need to set this in the config file.

See this for more detail on the switch to NodeJS:
https://github.com/facebook/jest/pull/9874