Decouple expiry from alert publicity level
Previously we marked all "operator"-type alerts as expired, which
was actually a bit pointless because there was nothing to stop the
"is_current" method also returning True. In practice, we only added
test alerts after the fact, but that won't be true once we switch
to using an API, so we need to fix the bug now.

Whether or not we show an alert on a page should be separate to the
facts about it: an alert should only be "expired" if its expiry time
has passed. To fix the bug and make the code clearer, I've renamed
the two filter / scope properties in Alerts, but kept the predicate
logic in Alert - DRYing up the Alerts tests to match.

Note that the strange form of 'mocker.patch' in test_alert(s).py is
necessary in order to mock a property. While it's a little unusual,
I think it's worth it compared to loads of duplication.