Add missing tests for showing alerts
This should give us more confidence the current/past alerts pages
will show the correct set of alerts.

Note that the "setup_jinja_environment" function was being called
with the wrong type of data, which the Jinja rendering was silently
ignoring. We always need to pass an instance of the Alerts class,
especially here where we're mocking its methods.