DRY-up rendering HTML in tests
Previously we had a localised function to do this in one test file.
There's no value in typing the same three lines in every test. This
has the added benefit that all tests now pass even if the dist/ dir
is empty. If we do need to test file fingerprinting, we can always
turn it on for just the tests that need it.