Replace "message_type" with "channel"
Rather than forcing the API to learn about "message types", we can
change this app to use the existing concept of a channel to determine
whether a message is public.

Note that the very first alert was actually two alerts on different
channels; we'll fix that in future work and for now we can just use
either of the channels as they're not shown anywhere.