Rename and sync "sent" field with API / DB
We dispatch ("send") alerts as soon as they are approved, so this
field is the closest match to what we had before. For public alerts
we previously set the value of the field to several minutes before
the start time to ensure the alert was rendered properly. This has
no effect once the start time for the alert has passed, so we can
safely change it to match the actual approval time.