Support alerts that are not (yet) cancelled
This is a prerequisite for dynamic alerts. I've remove the value of
"cancelled_at" for the 3 non-DB alerts since we don't know the actual
value and it's helpful to have some sample data of what an alert that
wasn't (yet) cancelled will look like.

Note that the "finishes_at" timestamps are always exactly 4 hours after
the "starts_at" timestamp, since we populate them at the same time with
the same constant offset [1].

[1]: https://github.com/alphagov/notifications-admin/blob/276644f504bb8d86a1cb9d0028cffa288049f17c/app/models/broadcast_message.py#L221-L224