Filter alerts to ensure they're in the UK
While this isn't an issue for alerts added statically in the YAML
file, we want to have a clear method for doing this before we try
to swap it with a dynamic API. An alert outside the UK is likely
to be *very* far outside, so the bounding box is generous. I've
added in the polygons for one static alert to help illustrate what
this data will look like eventually with an API.

We need to filter the list of alerts early as it gets used for more
than just presentation e.g. the last updated timestamp.

The code in "from_yaml" isn't meant to be final and it's expected
we'll rewrite it as part of adding an API data source. The focus
for someone reviewing this should be the new "is_in_uk" method,
which takes the simplest approach I could think of.