Fix bugs in test_all_pages.py
The tests in test_all_pages.py had a few mistakes
in their code which meant they ran green
incorrectly. The mistakes were:

In

test_local_links_lead_to_existing_routes_in_pages_with_alerts,

The alerts data instance was being passed to the ninja
environment wrapped in a list, which doesn't
happen in the build.

In

test_all_pages_with_details_in_have_the_js_for_it,

The data used to make the alert added current or
past mutated the alert in a loop so all of it was
applied at the same time.

This fixes those issues.