Switch tests to request HTML from Flask app
Resolves: https://github.com/alphagov/notifications-govuk-alerts/pull/151#discussion_r683329801

This will help ensure the development environment keeps working and
implicitly tests the render and Jinja setup functions. Since we no
longer have a custom Jinja environment for the tests, we need to have
a couple of mocks to make the normal one work.

Since the Flask app is never run in Production, it's safe enough to
remove the conditional preventing the pages from rendering in other
environments like 'test'.