Add architecture doc
The diagram is the important part, while the text is more about the
history of the app (to help explain potential weirdnesses). In future
we may want to rewrite this to simply state the architecture and how
the app is organised, but the story format seems more appropriate as
we continue to make changes towards the architecture in the diagram.

Command to generate the diagrams [1]:

        cat arch.dot | docker run -i tsub/graph-easy

[1]: https://gist.github.com/benthorner/1901916c925e1b60cd14728c9aa5d332