Automate the planned tests page
At the moment, whenever there is a planned test we need to update and
deploy gov.uk/alerts manually

We also update the site manually to announce upcoming public tests, like
Reading.

This commit:
- automatically puts a banner on the homepage and updates the ‘Planned
  tests’ page whenever there is an operator test happening on a given
  day
- gives us a way of publishing details of upcoming public tests without
  having to directly edit HTML (it’s a YAML file)

I’ve even made it work in the case when there’s an upcoming public test,
and a current operator test, like we had in the run-up to Reading.

For operator tests the banner will stay up until the end of the day. I
think this is sensible since the test might not last for very long, and
there might be several in one day (which we treat as one). For this to
work in a completely automated way we will need to schedule republishing
of the site at 00:01 each day.