add get_url_for_alert function
this function figures out the url for a given alert, based on the
alert's order in the day. Alerts are sorted by starts_at and then by id
to split ties.

If it's the first (or only) public alert in a day, we return the date
formatted as '3-jun-2021'. If it's the second or later public alert in
the day we give it a 1-indexed suffix, so starting at '3-jun-2021-2' and
counting up from there