Remove times that start at midnight from data.yaml
We do this because in the next commit we are going to change all
the times from BST to UTC and this will move the hour back one. This
would mean for times that start at midnight, they would become 11pm
on the day before. Whilst this is technically fine, I think it will
make this file less confusing and error prone if we change them to
later in the day.

Note, 10am has been chosen fairly randomly. It probably is a somewhat
realistic time that these operator alerts were sent. But more importantly,
it doesn't really matter what time we put as we don't show this information
anywhere in the user interface.