Convert  all timestamps in data.yaml to naive UTC
We want this file to mimic the API, which sends its times as naive
UTC timestamps. This will mean we have a consistent time format
when receiving information about alerts, whether they be from
the API or from the data.yaml file.

Note, this commit only changes the timestamps in the file, which
will mean that all dates will be an hour earlier and require some
changes in the next commit to display them correctly.