Add client for fetching alerts from Notify API
This is based on the one we have in the Admin app [1]. Note that
we don't need a few things that the Admin client has, which means
we don't need to override the 'generate_headers' method:

- ROUTE_SECRET_KEY - this is optional and not necessary here [2].

- Request ID stuff - not applicable as this app doesn't handle
user requests, which have a Request ID set.

I also took the opportunity to tidy up some of the variables in
the PaaS manifest. See here [3] to see how the "client ID" and
"client secret" are related in the API app.

[1]: https://github.com/alphagov/notifications-admin/blob/ff32e73d9bda80e556a9ee81fb0d867eef35e1e2/app/notify_client/__init__.py
[2]: https://github.com/alphagov/notifications-admin/pull/1851
[3]: https://github.com/alphagov/notifications-api/blob/b8aea23f6a7360d0830f6b24034627979550c824/app/config.py#L407