Refactor Alerts to support API data source
We want to apply the "in UK" filter to alerts coming from the API
as well, even though they should have been filtered already. Moving
this up to the "load" method has a knock-on affect on the return
value for the from_yaml method, so I had to rewrite a few tests.

I've switched to testing against the real data.yaml file since we
now know it will always have at least one alert in it. I'm not sure
why we were testing "last_updated_date" in the test to load alerts
as it's covered lower down. Since it doesn't work with no current
alerts, I've removed the assertion.