Refactor all timestamps to be timezone-aware UTC
Previously we switched all timestamps to naive (timezone-*un*aware)
UTC [1], but it turns out this is inconsistent with our API [2]. As
we start to include API data we'll hit a bug because it's not possible
to compare timezone-aware and timezone-unaware timestamps. Since most
of our API timestamps are timezone-aware, it makes sense to leave them
as-is and change all the timestamps here to be timezone-aware.

Rather than going back to specifying 'tzinfo=pytz.utc' everywhere,
and in the spirit of being consistent with the API, I've gone with
a less verbose option of parsing ISO 8601 strings and have created a
convention of aliasing dateutils.parser.parse as dt_parse. On the whole
I think the end result is a bit easier to read than lots of arguments.

I've also added an extra test to enforce timestamps in data.yaml are
timezone-aware UTC in case future alerts are added.

[1]: https://github.com/alphagov/notifications-govuk-alerts/pull/192
[2]: https://github.com/alphagov/notifications-api/blob/c84daf0b7bdb5b348475571bf5dd81c04770886a/app/utils.py#L15