Parse timestamps in alerts API client
This matches the post-processing done by yaml.load for static alert
data. While we could do the parsing later on, the string timestamps
aren't useful until they're parsed. Since the API client has the
most knowledge of the API, I think it makes most sense to parse here.

It's possible that some timestamps will be blank e.g. if the broadcast
was never cancelled. Since the test fixture only supports None for
cancelled_at and since the functionality is generic, I've only tested
for this field - that should be enough to minimise the risk of a bug.