Extend publish recipe to check it was successful
This repeats an approach we have to check API migrations have been
successful [1][2]. Rather than split the code across the Makefile
and a script in Concourse, I thought it was worth the effort to
make the existing recipe behave synchronously. This means we don't
need to make any changes to Concourse, but does mean the script is
a bit more compact than the one we have for API. Hopefully all the
comments are enough to make it maintainable in the future.

[1]: https://github.com/alphagov/notifications-api/blob/c84daf0b7bdb5b348475571bf5dd81c04770886a/Makefile#L159
[2]: https://github.com/alphagov/notifications-aws/blob/fc34c0493b60887d32d2aa9fef67dbaa18eb6e7f/concourse/templates/api.yml.j2#L150-L166