Hide content and areas of operator tests
We want to always show a standard message for operator tests on the
planned tests page, not the content of the alert itself.

We also don’t want to show the names of the areas targeted by an alert,
because past experience has shown that people interpret this to mean
that they are more likely to receive the test than they actually are.